backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "todo-backend-api"
version = "0.1.0"
description = "FastAPI backend for the single-user Todo web application (without authentication)"
authors = [{name = "Aliyan Aqeel", email = "developer@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.1",
    "sqlmodel>=0.0.16",
    "uvicorn>=0.24.0",
    "asyncpg==0.30.0",
    "pydantic>=2.10.3",
    "python-multipart>=0.0.6",
    "alembic>=1.13.1",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "email-validator>=2.1.0",
    "bcrypt>=5.0.0",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "httpx>=0.28.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "httpx>=0.25.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]
